# Fix DOM tree traversal optimization