Tree traversal is currently inaccurate and leads to slight degradation of performance:
* $3.5s$ -> $4s$ for small input
* $15s$ ->  $16s$ for large input

This pull reworks the algo to grease the speed and reduce file sizes a little.
These improvements are likely because leaves are more accurately detected and sorted.